Make population of column source metadata in result sets optional
Summary:
This is a relatively clean application of https://github.com/twitter/mysql/commit/f56acd7193122744fa28ebe07847e8d35657c336
Details can be found in the commit message.

There is an important caveat worth repeating:

The minimal option should be used with care as it constitutes a
protocol violation that might affect certain connectors. For
example, the minimal option makes the JDBC Driver for MySQL
(Connector/J) remap fields of type BLOB to VARCHAR as it considers
fields with an empty table name as originating from functions. Also,
certain methods of JDBC classes such as ResultSetMetaData and others
that represent meta-information might not yield meaningful results
if the minimal option is set.

Test Plan: mtr

Reviewers: anca, drilibo

Reviewed By: drilibo

Subscribers: webscalesql-eng

Differential Revision: https://reviews.facebook.net/D60015
Differential Revision: https://reviews.facebook.net/D61797