Port v5.1 SQL_NO_FCACHE to 5.6
Summary:
When running full table scans, provide a mechanism to disable
caching in flash so that the working set is not evicted.  This is
particularly useful when doing database dumps and not blowing out the
entire working set for data that most likely will not be read again.
This should also reduce the overall writes performed to the flash,
improving endurance.

Adds Innodb_dblwr_page_number to SHOW GLOBAL STATUS.  This is the page
number at which the doublewrite buffer starts in the system tablespace.
We need this for flashcache.

We turn off caching by default in flashcache and only enable it when a
process asks for it.  As long as the process group has flashcache
enabled, then all members will have it enabled.

Internal:
This is a port of the entirety of all of the following mysql-5138 diffs:
	https://phabricator.fb.com/D99125
	https://phabricator.fb.com/D99147
	https://phabricator.fb.com/D121847
	https://phabricator.fb.com/D134052
	https://phabricator.fb.com/D141875
	https://phabricator.fb.com/D215452

As well as parts of this one:
	https://phabricator.fb.com/D137676

And just some basic refactoring, code-cleanup, and testing.

Test Plan:
Setup a slave and did the following:
1. SELECT * FROM TABLE INTO OUTFILE 'foo' saw writes to flash.
2. SELECT SQL_NO_FCACHE * FROM TABLE2 INTO OUTFILE 'foo2' and saw no
   writes to flash.
3. SELECT /*!40001 SQL_NO_FCACHE */ * FROM TABLE3 INTO OUTFILE 'foo2'
   and saw no writes to flash.
4. SELECT /*!40001 SQL_NO_CACHE */ /*!40002 SQL_NO_FCACHE */ * FROM
   TABLE4 INTO OUTFILE 'foo2' and saw no writes to flash.
5. Ran mysqldump on a host which supports SQL_NO_FCACHE and it auto
   detected properly
6. Ran mysqldump on a host which does not support SQL_NO_FCACHE and it
   does not add the comment hint to the SQL statement.

Reviewers: chip, mcallaghan, jtolmer, santoshb

Reviewed By: chip, jtolmer, santoshb

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D438167
Differential Revision: https://phabricator.fb.com/D768680
Differential Revision: https://phabricator.fb.com/D960364

Task ID: 963630, 2843108

Conflicts:
	storage/innobase/handler/ha_innodb.cc
	storage/innobase/include/trx0trx.h
	storage/innobase/trx/trx0trx.cc