Port v5.1 SQL_NO_FCACHE to 5.6
Summary:
When running full table scans, provide a mechanism to disable
caching in flash so that the working set is not evicted.  This is
particularly useful when doing database dumps and not blowing out the
entire working set for data that most likely will not be read again.
This should also reduce the overall writes performed to the flash,
improving endurance.

Adds Innodb_dblwr_page_number to SHOW GLOBAL STATUS.  This is the page
number at which the doublewrite buffer starts in the system tablespace.
We need this for flashcache.

We turn off caching by default in flashcache and only enable it when a
process asks for it.  As long as the process group has flashcache
enabled, then all members will have it enabled.

Test Plan:
Setup a slave and did the following:
1. SELECT * FROM TABLE INTO OUTFILE 'foo' saw writes to flash.
2. SELECT SQL_NO_FCACHE * FROM TABLE2 INTO OUTFILE 'foo2' and saw no
   writes to flash.
3. SELECT /*!40001 SQL_NO_FCACHE */ * FROM TABLE3 INTO OUTFILE 'foo2'
   and saw no writes to flash.
4. SELECT /*!40001 SQL_NO_CACHE */ /*!40002 SQL_NO_FCACHE */ * FROM
   TABLE4 INTO OUTFILE 'foo2' and saw no writes to flash.
5. Ran mysqldump on a host which supports SQL_NO_FCACHE and it auto
   detected properly
6. Ran mysqldump on a host which does not support SQL_NO_FCACHE and it
   does not add the comment hint to the SQL statement.

Reviewers: chip, mcallaghan

Reviewed By: chip