Port v5.1 process_can_disable_bin_log
Summary:
SUPER has been required to do 'set sql_log_bin=...'.
This provides an option that PROCESS is sufficient,
but only when process_can_disable_bin_log is set.

Internal:
This is a full port of this v5.1 patch:
	https://phabricator.fb.com/D165078

This is a direct port, but a lot changed, so it's mostly a re-write.
Also, test was renamed to tell sys_vars.all_vars that it was tested.

Test Plan: New test included, passes.

Reviewers: chip

Reviewed By: chip

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D606123

Task ID: 1817166