Fix our acl damage in mysql_upgrade
Summary:
For Admission Control, we added two columns to the acl table.
The code to read this in mysql 5.1 and 5.6 is completely terrible.
So, this causes 5.6 to explode.

Reading this is worked around in our version of 5.6 already,
but to get mysql_upgrade to add the three new 5.6 acl columns,
these two obsolete Admission Control columns must first be dropped.

So, this does that in the course of mysql_upgrade.

For this to work, mysql_upgrade also has to ignore:
	ERROR 1091: Can't DROP '<column_name>'; ...
So, this was also added to its expected errors list.

This, in turn, required updating of some perfschema test results,
since they include line numbers of errors.

Test Plan:
Deployed 5.1 and 5.6 RPMs on a spare, and upgraded cleanly!
Jenkins 'arc unit' likes this too.

Reviewers: jtolmer

Reviewed By: jtolmer

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D659565