Port v5.1 extra page stats for innochecksum
Summary:

Add #leaf_pages, #recs_per_page, #data_bytes_per_page, and
page_size_histogram to innochecksum.

Add per-page details for non-index pages as well, as previous code
prints per-page details only for index pages.

Internal:
The first is a full port of this v5.6 patch:
	https://phabricator.fb.com/D533419.
This is a direct code port.

This second is a full port of this v5.6 patch:
	https://phabricator.fb.com/D531168.
This is a direct code port.

Test Plan: Tested on Jenkins Server

Reviewers: steaphan

Reviewed By: steaphan

Differential Revision: https://phabricator.fb.com/D725873

Task ID: 2098511