Replication IO thread retries more frequently
Summary:
Change get_master_version_and_clock so that it returns a retryable error
code when queries to the master fail, regardless of the failure. It will
still return a non-retryable error if the master query succeeds but
returns data the slave views as bad.

Internal:
This is a direct port of: https://phabricator.fb.com/D501640

Test Plan: mtr

Reviewers: steaphan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D623744

Task ID: 1817710