Port v5.1 Extra Stats: Flushed Neighbor Counters
Summary:
Add counters for flushed neighbor pages.

Innodb writes dirty pages from the same extent when one of the pages in
the extent is eligible to be written. This is done to reduce disk seeks.
The pages in the same extent are called neighbor pages.

Counters in SHOW STATUS are exported for this:

Innodb_buffer_pool_neighbors_flushed_from_list
	Number of neighbor pages flushed from the buffer pool flush list.

Innodb_buffer_pool_neighbors_flushed_from_lru
	Number of neighbor pages flushed from the buffer pool LRU list.

This is displayed in SHOW INNODB STATUS:
	Neighbor pages flushed: 644 from list, 0 from LRU

Test Plan: Jenkins.

Reviewers: chip

Reviewed By: chip