try to fix slow_log_extra_*
Summary:
long_query_time is a session variable as well as a global variable.

The initial value of long_query_time = 1 is not long enough during parallel executions
so some of the queries we don't want to log occasionally shows up in the log file. So
use session varible to avoid this in slow_log_extra_blg.

There's no good way to avoid the error showed up in slow_log_extra_rows_examined_exceed
cause that was due to the slow log is not flushed to disk. If there's anyway to wait for
that event to happen, please let me know. I made the test more reasonable anyways.

Checks for COM_QUIT in sql_parse.cc since I don't think those are supposed to show up in log.

Test Plan: mysqltest.sh

Reviewers: chip, jtolmer

Reviewed By: jtolmer

Differential Revision: https://phabricator.fb.com/D631537

Task ID: 1872198