Fix an inifinite loop in buf_page_get_gen() while reading compressed pages
Summary:
After loading a compressed page in buf_page_get_gen() we
allocate a new block for decompression. The problem is that the compressed
page is neither buffer-fixed nor I/O-fixed by the time we call
buf_LRU_get_free_block(), so it may end up being evicted and returned
back as a new block.

We then call buf_page_hash_get() to see if the page hash has been
modified, and indeed it returns NULL because the original page in not in
the buffer pool anymore. In which case we go back to read the page
again causing an infinite loop.

Details of the bug is in here http://bugs.mysql.com/bug.php?id=61132.
The bug is fixed as same as described in
http://bugs.mysql.com/bug.php?id=61132. First the block is buffer-fixed so
that it cannot be evicted or relocated.

The previous diff are
https://phabricator.fb.com/D244673
https://phabricator.fb.com/D253529

Test Plan: arc unit, mtr.

Reviewers: jtolmer, nponnekanti, steaphan

Reviewed By: jtolmer

CC: db-eng@lists, akotulski

Differential Revision: https://phabricator.fb.com/D639897

Task ID: 1817216