Port Compression: Reduce Calls to btr_cur_optimistic_insert to 5.6
Summary:
During the record insertion, InnoDB finds out the page the record
should be inserted, acquires the latch for that page and attempts
to insert the record to the page. If there is not enough room in
the page then the b-tree structure must be changed so it releases
the page latch and then gets the b-tree latch. It then re-computes
the page to which the record should be inserted. Most of the time
this second page is the same page as the previously found page and
it still doesn't have room for the record insertion. Rarely,
another thread (eg. purge thread) gets a hold of the previously
found page and deletes some records on that page so that there is
room for insertion in InnoDB's second attempt.

This diff eliminates the unnecessary attempts to insert the record
into the page by checking if the page has been modified since the
last time the insertion had been attempted.

Internal:
porting from 5.6.9 to 5.6.10, also fixing the problem described below:

Original porting (5.1 to 5.6) message below:
--------------------------------------------------------------------------
Port the following mysql-5.1 changes to mysql-5.6:

The function row_ins_index_entry_low is separeted into two
implementations for clustered index and secondary index as
row_ins_clust_index_entry_low and row_ins_sec_index_entry_low.

The original porting only ported the changed to
row_ins_clust_index_entry_low but not the other function. Complete the
implementation in this change.

Not needed:
  * Wed Jun 15 17:24:39 2011 (10205477d5c9af32824c1e08d22f9b41593c1629) nizamordulu:  remove the unnecessary calls to btr_cur_optimistic_insert()
  * Tue Aug 21 17:32:19 2012 (7eb65b2a650e39e4d9c9b371bd7420cb3329fea4) nizamordulu:  Revert "remove the unnecessary calls to btr_cur_optimistic_insert()"

Actually ported:
  * Thu Aug 23 17:35:16 2012 (cc60e4dda69bd791f16e300bb7c099480f912322) nizamordulu:  Reduce the number of calls to btr_cur_optimistic_insert()
       https://phabricator.fb.com/D557655

Change the upper limit count in innodb_optimistic_insert_race test.

There seems to be more calls to optimistic_inert in secondary index from
the pessimistic descend. The reason should be that secondary index
updates is more random thus has a higher chance of getting a different
page_no and modify_clock on the pessimistic insert call path from the
ones in optimistic insert. Experiments show that number of calls to
optimistic_insert in pessimistic descend from secondary index is likely
to be 3-4 times more than that of the clustered index in the given test.
Thus we increase the upper limit of number of calls allowed. There are
around 5000 calls to pessimistic insert from clustered index and
secondary index total. 300 is still a small number compare to that.

Test Plan:
innodb.innodb_optimistic_insert_race executes two threads:
an inserter and a deleter. The inserter thread inserts records in random
order while the deleter deletes in random order. There is a debug-only
status variable called
'innodb_optimistic_insert_calls_in_pessimistic_descent' which counts the
number of times btr_cur_optimistic_insert() needed to be called during
the pessimistic descent. This number is around 13 for the numbers
hardcoded in this test which is much less than the number of total
pessimistic descents (which is in thousands).

Reviewers: nizamordulu

Reviewed By: tianx
(Original Reviewed By: nizamordulu)

CC:

Differential Revision: https://phabricator.fb.com/D640596
Differential Revision: https://phabricator.fb.com/D644518
Differential Revision: https://phabricator.fb.com/D715332
Differential Revision: https://phabricator.fb.com/D720658

Task ID: 2121244