Port http://bugs.mysql.com/60682 to 5.6
Summary:
Fix http://bugs.mysql.com/bug.php?id=60682 which is a deadlock
when InnoDB calls thd_security_context while holding kernel_mutex. The
change is to use pthread_mutex_trylock in thd_security_context and display
"::BUSY::" for the query text when the lock is not acquired.

Previous diff from which this is ported
https://phabricator.fb.com/D231060
https://phabricator.fb.com/D233517
The porting is straight forward with a change in mutex calls
and the sql command SHOW INNODB STATUS changed to
SHOW ENGINE INNODB STATUS

Test Plan: Added test with reproduction case from bug 60682

Reviewers: jtolmer, steaphan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D641933

Task ID: 1817214