Make DROP DATABASE replicate safely with FK constraints
Summary:
Currently if you issue a DROP DATABASE dbname, MySQL will
actually log to the binary log a multi-table DROP TABLE statement.

If a database consists of enough tables, this DROP TABLE statement can
be split into 2 or more multi-table statements.

If a table in the first multi-table DROP TABLE statement has a FK
constraint that declares it a parent of a table not included in the DROP
TABLE statement, then when a replication slave or binary log playback of
the event occurs, MySQL will reject the DROP TABLE statement saying that
it would violate the FK constraint.

This bug can be avoided if the DROP TABLE statements are in a correct order
(child tables are dropped before parent tables and all tables in a
circular reference should be in one statement) or setting
OPTION_NO_FOREIGN_KEY_CHECKS flag in Query_log_event in which case
FK references do not matter when deleting parent tables. I used the later
since the order do not really matter in slave as they were already deleted
in master.

Test Plan: Added a test which reproduces the replicaiton bug.

Reviewers: jtolmer, lachlan, chip

Reviewed By: jtolmer

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D642895

Task ID: 1180841