Fix error injection for facebook crcs.
Summary:
This should be part of:
	https://phabricator.fb.com/D433170

This replaces a "return false" with a "break", to a "return false".

There is no actual change in release build, but this allows the
debug-only error injection to work for the "facebook" crc setting.

Test Plan: Jenkins; mysqltest.sh innodb.innodb-wl5980-linux

Reviewers: jtolmer

Reviewed By: jtolmer

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D659354

Task ID: 1964148