Fix default checksum method, just to be sure.
Summary:
This should be part of:
	https://phabricator.fb.com/D611040

The default was changed in mysql definition, but not C++ definition.
This wasn't a problem, but I want to fix it anyway.

The previous commit also means I don't have to override this to use
"innodb" crcs anymore - and this was done in same original diff.
This removes that change too.

Test Plan: Jenkins; mysqltest.sh innodb.innodb-wl5980-linux

Reviewers: jtolmer

Reviewed By: jtolmer

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D659385

Task ID: 1964148