Port rpl_read_size to 5.6
Summary:
This adds the my.cnf variable "rpl_read_size" with a default
of 8kb. It must be a multiple of 4kb. This sets the size for reads
done from the relay log and binlog. Making it larger might help with
IO stalls while reading these files when they are not in the OS buffer
cache.

Ported the following changes to 5.6

https://phabricator.fb.com/D488197.

Test cases is modified to differentiate sys_var test.
Added missing --source include/rpl_end.inc in the test

Test Plan: Ported test case and added sys_var test.

Reviewers: jtolmer, steaphan

Reviewed By: jtolmer

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D653373

Task ID: 1817726