Add support for MYSQL_SYSVAR_DOUBLE.
Summary:
This diff adds support for MYSQL_SYSVAR_DOUBLE.

This is a partial port of D333501, and a full port of D429341, and D438444.

1. D333501: Fix http://bugs.mysql.com/62534 by making the my.cnf innodb_max_dirty_pages_pct variable a double.
Only added the support for MYSQL_SYSVAR_DOUBLE from this diff.

2. D429341: The struct used to represent a global variable option my_option in my_getopt.h, has a field of type longlong to represent the value for scalar variables. This truncates the double values to integer if the global variable is not defined in my.cnf.

This currently affects us. innodb_comp_fail_max_fail_rate's default value is 0.1 but it is set to 0 in production because of this bug.

3. D438444: Fix the bug that divided the default values for double sysvars but did not multiply them.

Test Plan: Tested by creating a double sysvar and reading it in a test file. Also, tested on Jenkins.

Reviewers: nizamordulu, jtolmer

Reviewed By: jtolmer

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D740454

Task ID: 1957313