Add my.cnf var reset_seconds_behind_master and SHOW STATUS value Relay_log_io_connected
Summary:
When TRUE (default value) reset_seconds_behind_master preserves original
behavior which is to reset Seconds_Behind_Master to zero when the SQL
thread catches up to the IO thread.

When FALSE the reset is not done to avoid SBM flip-flops between zero and
the real lag.

The SHOW STATUS value Relay_log_io_connected counts the number of times the IO
thread connects to the master. This might detect network problems from frequent
disconnects.

This diff is a port of all or part of the following diffs from the 5.1 branch:
  e528a1ee886b6ea7bd567777672e98fc00a54c5b
  78294376f04d4c7c9847533e6bf502add5c74944
  c808e3ec1c6dc18952479b8b3754445b6ae0302d

Changes from those diffs are that
   rpl_sync_relay.test is not being ported so that portion of 78294376f0 was
     ignored.

   Parallel replication support means that there was an additional location
     which was resetting last_master_timestamp where I needed to add a check
     for reset_seconds_behind_master.

   I had to change the logic in show_slave_status which populates the field
     to not force 0 if !reset_seconds_behind_master.

   There were additional tests changes vs 5.1 because of the new sys_vars suite.

Test Plan: mtr

Reviewers: steaphan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D663325

Task ID: 1817692

Conflicts:
	mysql-test/r/mysqld--help-notwin.result