Test autoinc intvar not persisted across relay-log events on slave
Summary:
Adds a test case to check autoinc intvar are not persisted across
relay-log events on slave.

When master and slave have different schemas, in particular different
AUTO_INCREMENT columns, INSERT_ID events logged for a given table on
master may be applied to a different table on slave on SBR. For example
in the following sequence:

1) Insert into table t3 created only in slave with auto-inc column.
2) Insert into table t1 with auto-inc column in master but not in
   slave.
3) Insert or update into table t2 with no auto-inc column and has a
   trigger only in slave but not in master. The trigger inserts a row into
   table t3.
4) The insert into table t3 gets INSERT_ID from the previous insert into table
   t1 causing duplicate key errors.

The update on t2 causes inconsistency since the INSERT_ID values from
previous event (insert into t1) was used for inserting into table t3.
This is fixed in 5.6.10 branch. Added a test case to confirm.

Test Plan: mtr

Reviewers: jtolmer

Reviewed By: jtolmer