Change the upper limit count in innodb_optimistic_insert_race test.
Summary:
There seems to be more calls to optimistic_inert in secondary index
from the pessimistic descend. The reason should be that secondary index
updates is more random thus has a higher chance of getting a different
page_no and modify_clock on the pessimistic insert call path from the ones
in optimistic insert. Experiments show that number of calls to optimistic_insert
in pessimistic descend from secondary index is likely to be 3-4 times more than
that of the clustered index in the given test. Thus we increase the upper limit
of number of calls allowed. There are around 5000 calls to pessimistic insert
from clustered index and secondary index total. 300 is still a small number compare
to that.

Test Plan: mysqltest.sh

Reviewers: nizamordulu

Reviewed By: nizamordulu

Differential Revision: https://phabricator.fb.com/D720658

Task ID: 2121244