Set Default table_open_cache_instances=8
Summary:
Based on Mark Callaghan's recommendations, change this default:
	table_open_cache_instances: 1 -> 8

Also fix tests or results to account for the change.
3 of the tests I fixed in a way that they will continue to pass
despite the value of table-open-cache-instances with which
the server is running. Test table_open_cache_instances_basic
is hardcoded to the new default value.

For rpl_trigger, there is a new warning generated when executing
the SELECT from t1. The source of the warning is because the test
replays an old binlog where the trigger did not define a definer.
The test didn't previous generate a warning because the warning is
only generated when opening a table. Previously in the test the
replication SQL thread generated the warning, which isn't visible
in any way, but the SELECT did not because the table was served
out of the open table cache. When the cache instances > 1, the
SELECT hits a different cache, does not find the table and has
to reopen it. Thus, the warning is generated. I've updated the test
so that the warning is also generated for the cache instances == 1
by flushing the cache before the SELECT.

Test Plan: mtr

Reviewers: steaphan, mcallaghan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D715621

Task ID: 2115429