Add Facebooks's integrated version of xtrabackup
Summary:
This began with just the "innodb56.patch" from Percona, with some
fixes from them as well, including:

* https://bugs.launchpad.net/percona-xtrabackup/+bug/950334
* A fix for deleted tablespaces in xtrabackup apply log (from 5.1 version)

But, this is all done differently than that patch.  Instead of a
separate patch, this is integrated into the main code.

All changes are now protected via the XTRABACKUP define.

This allows the same code to compile for both mysqld and xtrabackup.

This should have a net ZERO effect, until the xtrabackup build is added.

We build this in our tree, beside our normal out-of-dir cmake build.

Basically, this can be built just by running this on a clean source tree:

	cd xtrabackup; ./utils/build.sh innodb56; cd ..

Note that this only works right in a clean source tree.
But, since we build out-of-dir, our source tree is always clean.

Also note that some corner-cases (Debug builds of xtrabackup, for example)
will probably still require more build magic to work.  The real solution
here would be to integrate this build into the main cmake build of
mysql-5.6 itself, but we didn't go that far with this (yet).

Sections not required for mysql-5.6 were removed, for simplicity.

Also, some compile/build warnings/errors were cleaned up.

And, innobackupex was modified.  For apply log, if xtrabackup binary is
explicitly specified, dont try to autodetect.

Also, we initialize error messages for xtrabackup, and we do not use
thread-specific locales for error messages.

Also, we create the xtrabackup_logfile in the supplied target_dir.

Also, we removed the FLUSH TABLES WITH READ LOCK calls from innobackupex.

Also, we added a fix to prevent an indefinite sleep of xtrabackup.

Also, we added support for the O_DIRECT_NO_FSYNC option.

And, we don't use O_DIRECT for xtrabackup_logfile.

Test Plan:
We created a small suite of xtrabackup tests that run within mtr.
These are included here, and all pass.

Jenkins (arc unit).  All pass.

xtrabackup build works (in CLEAN source dir):
	cd xtrabackup; ./utils/build.sh innodb56; cd ..

Artificially threw some warnings in xtrabackup.
Successfully carried out a full backup with warning (segfaulted before).

Reviewers: steaphan, jtolmer, nizamordulu

Reviewed By: steaphan, jtolmer, nizamordulu

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D722664
Differential Revision: https://phabricator.fb.com/D733605
Differential Revision: https://phabricator.fb.com/D734848
Differential Revision: https://phabricator.fb.com/D736104
Differential Revision: https://phabricator.fb.com/D737487
Differential Revision: https://phabricator.fb.com/D738773
Differential Revision: https://phabricator.fb.com/D740503
Differential Revision: https://phabricator.fb.com/D740640
Differential Revision: https://phabricator.fb.com/D741003
Differential Revision: https://phabricator.fb.com/D746758
Differential Revision: https://phabricator.fb.com/D748307
Differential Revision: https://phabricator.fb.com/D754239
Differential Revision: https://phabricator.fb.com/D757601
Differential Revision: https://phabricator.fb.com/D781326
Differential Revision: https://phabricator.fb.com/D808542
Differential Revision: https://phabricator.fb.com/D806754
Differential Revision: https://phabricator.fb.com/D810029
Differential Revision: https://phabricator.fb.com/D810101
Differential Revision: https://phabricator.fb.com/D820339
Differential Revision: https://phabricator.fb.com/D824025
Differential Revision: https://phabricator.fb.com/D825203
Differential Revision: https://phabricator.fb.com/D829066
Differential Revision: https://phabricator.fb.com/D830852
Differential Revision: https://phabricator.fb.com/D834046
Differential Revision: https://phabricator.fb.com/D836340
Differential Revision: https://phabricator.fb.com/D846538
Differential Revision: https://phabricator.fb.com/D864057
Differential Revision: https://phabricator.fb.com/D866262
Differential Revision: https://phabricator.fb.com/D899565
Differential Revision: https://phabricator.fb.com/D921854
Differential Revision: https://phabricator.fb.com/D965930
Differential Revision: https://phabricator.fb.com/D967141
Differential Revision: https://phabricator.fb.com/D1036818
Differential Revision: https://phabricator.fb.com/D1052819
Differential Revision: https://phabricator.fb.com/D1055876
Differential Revision: https://phabricator.fb.com/D1108029
Differential Revision: https://phabricator.fb.com/D1177689

Task ID: 3086129, 2190294, 2151092, 2309092, 2715544, 2378610, 2380044, 2382338, 2285646, 2285646, 2285646, 2457004, 2457004, 2285646, 2650168, 3103060, 2865619, 2862432, 2862432