Adding logging for durability changes
Summary:
Adds logging for changes to sync_binlog and innodb_flush_log_at_trx_commit.

The code for sync_binlog was basically copy and paste from the similar
change to the read_only variable. In sys_vars.cc the change for
sync_binlog required that Sys_sync_binlog_period be extended to use more
of the available parameters in order to be able to use ON_UPDATE.

* The first new parameter notates which mutex protects changes to the
  variable. If no mutex is to be used (as in this case) the parameter
  should be NO_MUTEX_GUARD.

* The second new parameter designates whether a change to the variable
  should go into the binary log. In this case the change should not be
  logged.

* The third new parameter designates which function is to be called when
  the variables is accessed. In this case the function pointer is NULL.

* The fourth and final new variables is a function pointer to a new
  function log_sync_binlog_change which actually logs the changes of
  this variable.

Unfortunately the code where innodb_flush_log_at_trx_commit is updated,
inside of InnoDB, does not have easy access to the functions for dealing
a thread handler. This patch includes a bunch of glue for InnoDB to get
back into MySQL to gain information about the user and host.

Test Plan: new tests added to confirm sanity and they pass

Reviewers: steaphan, rongrong, jtolmer, mcallaghan

Reviewed By: steaphan