Implement START TRANSACTION WITH CONSISTENT INNODB SNAPSHOT
Summary:
Provide better alternative to FLUSH TABLES WITH READ LOCK
for use to take backups. See
mysqlha.blogspot.com/2009/10/be-careful-with-flush-tables-with-read.html
for context. This diff adds START TRANSACTION WITH CONSISTENT
INNODB SNAPSHOT which starts an InnoDB transaction and returns
the master binlog filename and offset which corresponds to all
transactions visible to the InnoDB view for the transaction.
When this is used for backups, replication should be restarted
at those return values. The new START command returns 1 row with
2 columns (File, Position) on success.

In order to get the correct binlog filename and position for the
view it is required to block all commit progress, which is
accomplished by acquring all the locks which are also acquired
by MYSQL_BIN_LOG::ordered_commit.

Test Plan:
Added new mtr tests. I also plan to add a stress test, which tests this
feature in combination with binlog group commits happening concurrently,
in a follow-up diff.

Reviewers: steaphan, hfisk

Reviewed By: steaphan