Add innodb_page_cleaner_interval_millis
Summary:
This adds a my.cnf option to set the interval at which the
page cleaner thread runs. The default is 1000 millis and the range
is 100 to 1000. The thread runs at most once per this many millis.
Setting it to a value less than 1000 allows smaller values to be
used for innodb_io_capacity and innodb_lru_scan_depth and hopefully
reduce mutex contention. Note that the page cleaner thread is
responsible for flushing dirty pages from the flush and LRU lists
and innodb_io_capacity determines how many pages can be flushed
from the flush list, innodb_lru_scan_depth does the same for the
LRU list. Those limits are per buffer pool instance, not global.

This has results for sysbench where a query is update 1 row by PK.
The database is ~32gb and the buffer pool is 4gb. The host has flash
and the test is IO bound. The results are for MySQL 5.6 with a few
patches from Oracle to fix stalls.

io=X means that innodb_io_capacity and innodb_lru_scan_depth=X

-- results without my change
Updates/second
    8    16    32    64   128   256 concurrent clients
15192 20771 24702 24643 24519 24516 fil_flush patch,O_DIRECT, io=1024
17231 23661 22663 21892 21230 21401 fil_flush patch,O_DIRECT, io=8192

--- results with my change and page_cleaner_interval_millis=125
Updates/second
    8    16    32    64   128   256    concurrent clients
17697 24218 26822 27034 27014 27020 fil_flush patch,O_DIRECT, io=1024
17486 24316 26651 26721 26493 26719 fil_flush patch,O_DIRECT, io=2048
17575 24035 25749 25505 25439 25231 fil_flush patch,O_DIRECT, io=4096
17211 23500 25157 24080 23812 24408 fil_flush patch,O_DIRECT, io=8192

Test Plan:
mtr

Reviewers: steaphan

Reviewed By: steaphan

CC: db-eng@lists

Differential Revision: https://phabricator.fb.com/D744746